build with a Go toolchain patched by Termux, and delete two of our patches
The interface-discovery and DNS problems this project works around are not
tailscale problems, they are Go standard-library problems on Android -- and
Termux already fixes both in the Go it ships, and applies the same patches when
cross-compiling other packages. Doing the same here replaces ~470 lines of our
own code with four vendored diffs.

Measured on an Android 16 phone in a real untrusted_app context, running the
same probe built two ways:

  vanilla Go:  net.Interfaces(): err=netlinkrib: permission denied  count=0
               LookupHost: ... on [::1]:53 ... connection refused
  patched Go:  net.Interfaces(): err=<nil>  count=8
               LookupHost: err=<nil>  ips=[...]

The patched build sees more than our patch did: cellular rmnet interfaces and
a global IPv6 address, which the ifconfig/procfs ladder never reliably got.
Then verified with the real daemon on the same phone -- registered with the
control plane, and `curl --socks5-hostname` through it returned 200, which is
the DNS path that matters.

So patches/fix_android_netmon.go is gone, along with the tsdial resolver
injection and the wlynxg/anet dependency. What is left of that file is the
hostinfo hook, which is a deliberate choice rather than a workaround, moved to
patches/fix_hostinfo_android.go.

Consequences worth stating:

- Every architecture is now GOOS=android with -buildmode=pie, so all four ship
  as ET_DYN against /system/bin/linker instead of one PIE and three ET_EXEC.
  That also covers Termux builds targeting API 29+, which launch binaries
  through the linker and reject ET_EXEC.
- Cross-compiling arm/i686/x86_64 now needs the NDK, because Go refuses
  GOOS=android without cgo there. Building on a phone needs nothing extra:
  Termux's own Go already carries the patches, so build.sh detects that and
  skips the toolchain download entirely.
- GOTOOLCHAIN=local is not optional. Without it Go silently downloads the
  toolchain named in tailscale's go.mod and discards every patch above.
- DNS is now whatever $PREFIX/etc/resolv.conf says -- the same resolver as the
  rest of Termux, editable by the user -- so TS_DNS_SERVER is gone and nothing
  is hardcoded to a public resolver. resolv-conf became a package dependency,
  since without that file there is no resolver at all.
- verify_patched checks the artifact for the toolchain's resolv.conf path: a
  toolchain patch that stops applying otherwise yields a binary that looks fine
  and cannot resolve a name or see an interface.
- Patch application is fatal on failure. These diffs patch Go's own sources and
  are sensitive to its version; they applied cleanly to go1.27.1.
