+++ patches/go/fix-android-netlink.diff
@@ -1,0 +1,226 @@
+https://github.com/golang/go/issues/40569
+https://developer.android.com/training/articles/user-data-ids#mac-11-plus
+https://github.com/wlynxg/anet
+
+net, syscall: fix interface enumeration on Android 11+
+
+Since Android 11, ordinary applications are denied bind(2) on netlink
+sockets and RTM_GETLINK dumps, so net.Interfaces fails with
+"netlinkrib: permission denied". Fall back to an RTM_GETADDR dump
+plus SIOCGIF* ioctls, as github.com/wlynxg/anet does.
+
+Fixes #40569
+
+--- a/src/syscall/netlink_linux.go
++++ b/src/syscall/netlink_linux.go
+@@ -2,6 +2,8 @@
+ // Use of this source code is governed by a BSD-style
+ // license that can be found in the LICENSE file.
+ 
++//go:build linux && !android
++
+ // Netlink sockets and messages
+ 
+ package syscall
+--- a/src/syscall/netlink_android.go
++++ b/src/syscall/netlink_android.go
+@@ -2,6 +2,8 @@
+ // Use of this source code is governed by a BSD-style
+ // license that can be found in the LICENSE file.
+ 
++//go:build android
++
+ // Netlink sockets and messages
+ 
+ package syscall
+@@ -65,7 +67,11 @@
+ 	defer Close(s)
+ 	sa := &SockaddrNetlink{Family: AF_NETLINK}
+ 	if err := Bind(s, sa); err != nil {
+-		return nil, err
++		if err != EPERM && err != EACCES {
++			return nil, err
++		}
++		// Android 11+ denies bind(2) on netlink sockets; the kernel
++		// assigns the port ID of an unbound socket on the first send.
+ 	}
+ 	wb := newNetlinkRouteRequest(proto, 1, family)
+ 	if err := Sendto(s, wb, 0, sa); err != nil {
+--- a/src/net/interface_linux.go
++++ b/src/net/interface_linux.go
+@@ -2,6 +2,8 @@
+ // Use of this source code is governed by a BSD-style
+ // license that can be found in the LICENSE file.
+ 
++//go:build linux && !android
++
+ package net
+ 
+ import (
+--- a/src/net/interface_android.go
++++ b/src/net/interface_android.go
+@@ -2,9 +2,12 @@
+ // Use of this source code is governed by a BSD-style
+ // license that can be found in the LICENSE file.
+ 
++//go:build android
++
+ package net
+ 
+ import (
++	"bytes"
+ 	"os"
+ 	"syscall"
+ 	"unsafe"
+@@ -16,7 +19,9 @@
+ func interfaceTable(ifindex int) ([]Interface, error) {
+ 	tab, err := syscall.NetlinkRIB(syscall.RTM_GETLINK, syscall.AF_UNSPEC)
+ 	if err != nil {
+-		return nil, os.NewSyscallError("netlinkrib", err)
++		// RTM_GETLINK dumps are denied on Android 11+; fall back to
++		// RTM_GETADDR plus SIOCGIF* ioctls.
++		return fallbackLinkTable(ifindex)
+ 	}
+ 	msgs, err := syscall.ParseNetlinkMessage(tab)
+ 	if err != nil {
+@@ -45,6 +50,64 @@
+ 	return ift, nil
+ }
+ 
++// fallbackLinkTable enumerates interfaces via an RTM_GETADDR dump
++// plus SIOCGIF* ioctls when RTM_GETLINK is denied. Interfaces
++// without any address are not reported.
++func fallbackLinkTable(ifindex int) ([]Interface, error) {
++	tab, err := syscall.NetlinkRIB(syscall.RTM_GETADDR, syscall.AF_UNSPEC)
++	if err != nil {
++		return nil, os.NewSyscallError("netlinkrib", err)
++	}
++	msgs, err := syscall.ParseNetlinkMessage(tab)
++	if err != nil {
++		return nil, os.NewSyscallError("parsenetlinkmessage", err)
++	}
++
++	var ift []Interface
++	im := make(map[uint32]struct{})
++loop:
++	for _, m := range msgs {
++		switch m.Header.Type {
++		case syscall.NLMSG_DONE:
++			break loop
++		case syscall.RTM_NEWADDR:
++			ifam := (*syscall.IfAddrmsg)(unsafe.Pointer(&m.Data[0]))
++			if _, ok := im[ifam.Index]; ok {
++				continue
++			}
++			im[ifam.Index] = struct{}{}
++
++			if ifindex == 0 || ifindex == int(ifam.Index) {
++				ifi := newIoctlLink(ifam.Index)
++				if ifi != nil {
++					ift = append(ift, *ifi)
++				}
++				if ifindex == int(ifam.Index) {
++					break loop
++				}
++			}
++		}
++	}
++	return ift, nil
++}
++
++// newIoctlLink returns the Interface for the given index using
++// ioctls, or nil if its name cannot be determined.
++func newIoctlLink(index uint32) *Interface {
++	name, err := indexToName(index)
++	if err != nil {
++		return nil
++	}
++	ifi := &Interface{Index: int(index), Name: name}
++	if mtu, err := nameToMTU(name); err == nil {
++		ifi.MTU = mtu
++	}
++	if flags, err := nameToFlags(name); err == nil {
++		ifi.Flags = flags
++	}
++	return ifi
++}
++
+ const (
+ 	// See linux/if_arp.h.
+ 	// Note that Linux doesn't support IPv4 over IPv6 tunneling.
+@@ -255,3 +318,75 @@
+ 	}
+ 	return ifmat
+ }
++
++// ifReq is large enough to hold a struct ifreq.
++type ifReq [40]byte
++
++func ioctl(fd int, req uint, arg unsafe.Pointer) error {
++	_, _, errno := syscall.Syscall(syscall.SYS_IOCTL, uintptr(fd), uintptr(req), uintptr(arg))
++	if errno != 0 {
++		return errno
++	}
++	return nil
++}
++
++// indexToName returns the name of the interface with the given index.
++func indexToName(index uint32) (string, error) {
++	fd, err := syscall.Socket(syscall.AF_INET, syscall.SOCK_DGRAM|syscall.SOCK_CLOEXEC, 0)
++	if err != nil {
++		return "", err
++	}
++	defer syscall.Close(fd)
++
++	var ifr ifReq
++	*(*uint32)(unsafe.Pointer(&ifr[syscall.IFNAMSIZ])) = index
++	if err := ioctl(fd, syscall.SIOCGIFNAME, unsafe.Pointer(&ifr[0])); err != nil {
++		return "", err
++	}
++
++	return string(bytes.Trim(ifr[:syscall.IFNAMSIZ], "\x00")), nil
++}
++
++// nameToMTU returns the MTU of the interface with the given name.
++func nameToMTU(name string) (int, error) {
++	// Leave room for terminating NULL byte.
++	if len(name) >= syscall.IFNAMSIZ {
++		return -1, syscall.EINVAL
++	}
++
++	fd, err := syscall.Socket(syscall.AF_INET, syscall.SOCK_DGRAM|syscall.SOCK_CLOEXEC, 0)
++	if err != nil {
++		return -1, err
++	}
++	defer syscall.Close(fd)
++
++	var ifr ifReq
++	copy(ifr[:], name)
++	if err := ioctl(fd, syscall.SIOCGIFMTU, unsafe.Pointer(&ifr[0])); err != nil {
++		return -1, err
++	}
++
++	return int(*(*int32)(unsafe.Pointer(&ifr[syscall.IFNAMSIZ]))), nil
++}
++
++// nameToFlags returns the flags of the interface with the given name.
++func nameToFlags(name string) (Flags, error) {
++	// Leave room for terminating NULL byte.
++	if len(name) >= syscall.IFNAMSIZ {
++		return 0, syscall.EINVAL
++	}
++
++	fd, err := syscall.Socket(syscall.AF_INET, syscall.SOCK_DGRAM|syscall.SOCK_CLOEXEC, 0)
++	if err != nil {
++		return 0, err
++	}
++	defer syscall.Close(fd)
++
++	var ifr ifReq
++	copy(ifr[:], name)
++	if err := ioctl(fd, syscall.SIOCGIFFLAGS, unsafe.Pointer(&ifr[0])); err != nil {
++		return 0, err
++	}
++
++	return linkFlags(*(*uint32)(unsafe.Pointer(&ifr[syscall.IFNAMSIZ]))), nil
++}
